+++ src/utils/email.rs
@@ -6,17 +6,23 @@ use lettre::{
 
 use crate::EmailConfig;
 
+/// Email client.
 #[derive(Clone)]
 pub struct Email {
-    addr: String,
+    /// Mailbox to send email from.
     from: Mailbox,
+    /// Underlying SMTPS transport.
+    transport: SmtpTransport,
 }
 
 impl Email {
     pub async fn connect(config: EmailConfig) -> Result<Self> {
-        // we need this for smtps
+        // `lettre` requires a default provider to be installed to use SMTPS.
         let _ = rustls::crypto::aws_lc_rs::default_provider().install_default();
-        Ok(Self { addr: config.smtp_addr, from: config.from })
+
+        let transport = SmtpTransport::from_url(&config.smtp_addr)?.build();
+
+        Ok(Self { transport, from: config.from })
     }
 
     pub fn builder(&self) -> MessageBuilder {
@@ -24,8 +30,7 @@ impl Email {
     }
 
     pub async fn send(&self, message: Message) -> Result<()> {
-        let transport = SmtpTransport::from_url(&self.addr)?.build();
-        transport.send(&message)?;
+        self.transport.send(&message)?;
         Ok(())
     }
 }
