+++ src/utils/mod.rs
@@ -1,3 +1,6 @@
 pub mod config;
 pub mod db;
 pub mod email;
+pub mod tera;
+pub mod tracing;
+pub mod types;
